Move method to OriVertex.
